Refactor FindMutation class
This is part of the multi PR, which aim to refactor `call.cc`, FindMutation class, and MutationStats class.

Refactor FindMutation class out of call.cc.
Both constructor and operator() in find_mutation.cc are exactly the same as the previous version in `call.cc`.

The unittest for find_mutations.cc will be included in the next PR. Becasue it depend on some fixture in the other PR. So the unittest is skipped in this PR to avoid merging conflict.