Replace LLH with LLS (#163)
LLH is a statistic that contains the log-likelihood of the data scaled such that the most likely genotype of each library is 1.

LLS is a statistic that compares the log-likelihood of the data versus a maximum likelihood binomial model. This statistic can thus indicate how much better the model we use to call nucleotides does than the NULL, pooled binomial model. This allows us to control for total read depth.

Unlike LLH, LLS is interpretable and thus we will remove LLH and add LLS.


* scale and check nucleotide frequencies

* change how LLH is calcualted part 1

* Calcualte LLH based on alphas

* Calculate LLH by comparing LLD to the best fitting binomial

* Drop LLH in favor of LLS which compares LLH to a MLE binomial of the pooled read counts