Basic functionality for generating visualization
Invokes Rscript pedigree_and_layout.R to compute the layout
(depends on R kinship2 package), and bundles the output into
an html template. The final output is an html file which
contains all the visualization data except for some 3rd party
libraries like JQuery, d3, and Bootstrap, which are currently
being pulled from the web by the user's browser at runtime.

Note: this functionality depends on pedigree_and_layout.R and
the index.html template, which are not included in this commit,
since they are part of the dng-vis repo. I'll probably add them
in a separate commit.