Feature: Simplify Genotype Likelihoods (#225)
Using a mixture of 2 DM for genotype likelihoods ended up being too complicated from the standpoint of the end user. The gamma parameter was too complicated. The solution is to change the model to a single dirichlet multinomial, which allows the command line parameters to be simplified.

The benefit is that calculating genotype likelihoods becomes faster.

While creating unit tests, bugs in the calculation of haploid genotype likelihoods were discovered and fixed. Other tests were adjusted to reflect the new genotype likelihood model and default parameters.

This pull request also includes improvements to some additional unit tests that began on another branch, but needed to be merged into this one.

Squashed Comments
=============
* Improve some unit tests
* update get_high_bit unit_test
* Begin creating new peeling tests
* Correct compiling error when building unit tests
* Simplify genotype likelihood model to be a single DM. This speeds up calculations and makes commandline parameters less cryptic.
* Add unit tests for DirichletMultinomial
* Fix bugs revealed by tests

* Use the xorshift header for Unit.dng::io::ad

* Update genotype unit tests and fix bugs that were discovered

* Update Tests to reflect the new default genotype model

* Update CHECK_CLOSE_RANGES macro for Boost.Test 2 to better
mimic behavor of Boost.Test 3.

* Fix compiling error on Darwin